[bootstrap] Let's write a more feature-complete compiler in Rust
The approach we took in the previous attempts to get to a self hosted
version of the compiler as quick as possible, proved to be wrong.

We need more language features to write a basic self hosted compiler.
Furthermore, by writing a more complete compiler in Rust first, we learn
a lot about the approach we want to take when writing the self hosted
compiler.