typed+ir+codegen: Add IR register constraints
Now that we have else branches we need to make sure that symbol values
that are modified in both branches land in the same hardware register or
stack location.

Consider this simple if/else and its corresponding IR:

mut a = 1
if true {
    a = 2
} else {
    a = 3
}

_main:
    %1 = i64 1          -- mut a = 1
    %2 = i1 1           -- `true`
    %3 = condbr i1 %2, true_block, else_block
true_block:
    %4 = i64 2          -- a = 2
    jmp merge_block
false_block:
    %5 = i64 3          -- a = 3
    jmp merge_block
merge_block:
    !!! Here, `a` may be in %4 or %5 depending on which branch was taken.

Other compiler frameworks like LLVM insert a so-called phi node to
communicate that we have to look for the value depending where we came
from:

merge_block:
    %6 = phi %4 true_block, %5 else_block

This basically says: If you arrived here from the `true_block` then
%6 will have the value of %4, but %5 if we came from the `else_block`.

We are taking a slightly different approach (with the same result).
During IR generation we add a register constraint telling the code
generation to make sure to put %1, %4, and %5 into the same hardware
register or stack location.