codegen: Restore the caller saved registers after a call
We need to actually restore the caller saved registers after a function
call. For linear programs, i.e. ones without branches or loops, this was
not needed, we just would restore the value from the stack once it was
needed. But with branches and loops the values that use register
constraints we require the values to stay in the hardware registers.

Like most things we can of course optimize this behaviour, but first
let's concentrate on correctness and readable code. :-)