typed+lower+ir+codegen: Remove name mangling and use `TypeId` in codegen
We continue getting away from using names after the typechecker. With
this patch we removed the name mangling we did in the lowering step.

Now we use the `typed.TypeId` for symbols in the IR and codegen. This
makes the generated IR and ASM more difficult to read but we plan to add
debug symbols in a later step to mitigate this and have debug info in
the actual build.