typed: Introduce Symbols and Scopes while removing names from types
Types like function, method, trait, or struct should not contain their
names but only their signature. This makes future features like function
types (i.e. functions as first class citizens) much easier.

With the previous commit that removed name mangling in favor of using
the `TypeId` to name things in the IR and ASM we lost a bit of
debuggability. We introduces Symbols so that we can lookup names and
scopes in later stages like IR or ASM generation. We use symbols to add
a comment with the function name to ASM. Later on we plan to expand on
this and add full featured debug symbols to the binaries to make it
possible to run Klar programs in a debugger.