typed+lower+ir+codegen: Use proper `FunctionType` in IR
This patch cleans up the resolution of functions in call expressions. In
the IR we don't want to rely on names (that originated in the user's
source code) but only type ids. And function resolution was the last
place that relied on names.

Going forward we want to support function pointers (making functions
first class citizens) and indirect calls. This patch lays the
foundations for that.