ir+codegen: Clean up the use of `ir.Register`, `ir.Call`, and `ir.GetPointer`
Before this patch we put everything into ir.Register and then based on
the register value we would determine wether to load a pointer to a
string constant or a function or just a regular register. Same with how
we determined whether to make a direct or indirect call.

We now cleaned this up and have a proper ir.GetPointerKind and
ir.Callee.