codegen: Fix spilling of used registers
This patch changes the way the register allocator works. Instead of a
map of register to allocation we now use a ring buffer to store the used
allocations and which register they occupy. This has the effect that
always the oldest allocation is spilled and not a random one.

With the old implementation we ran into a bug where we would spill a
register we just allocated.