ast+typed+lower+ir: Move type arguments away from `ast.CallExpression`
... and into `ast.IdentExpression` and `ast.MemberExpression`. This
simplifies handling of generic types a lot and enables generics to be
used in more places like assignments:

    fn foo<T>(val T) T {
        val
    }
    let str_foo = foo<Str>
    print(str_foo("Hello, world!"))

The downside is that type inference is a bit more complex because we
have to keep some kind of "inference scope" while type checking whereas
formally we were able to infer the types directly in
`VisitCallExpression`. But we need the more complex type inference
anyway in the future, I guess.

And lastly, we detected that our function specialization
(monomorphization) is not correct, as it cannot detect specializations
needed in sub-scopes like this:

    fn foo<T>(val T) T { val }
    fn bar<U>(val U)  {
        foo<U>(val)
    }
    foo<Int>(42)
    bar<Str>("Hello")

We don't detect the specialization of `foo<Str>`. Currently this all
magically works, because we record the specialization `foo<U>` and treat
`U` as an opaque pointer type. This is obviously not correct, so we have
to fix it in a future patch.