ast+typed+lower+ir+codegen: Rewrite monomorphization
Our previous attempt at monomorphization was not complete and overly
complex.

This patch separates the AST transformation (`prepare.go`) from the actual
monomorphization (`monomorphize.go`) in the lowering pass.

We also returned back to using FQN in the IR and ASM instead of just
type-ids. This makes both much more readable.

Generic structs are not supported by this patch, but they weren't
before. :-)