typed+lower: Refactor: Add `typed.GenericsResolver`
This is the first part in a refactoring that creates new types (i.e.
with their own unique `TypeId`) for each new resolved generic type
during typechecking. Currently we just use the same `TypeId` as the base
generic type (struct or function) but with different `TypeArgs`. This
does not work with self-referential types (and other usages probably
too).