typed+ir: Add `sizeof<T>()`
This is in preparation of the manual memory management we are going to
implement to make it possible to write basic types like `Array` or `Map`
in Klar itself instead of implementing them in the compiler.

This manual memory management will only be available in the core library
of Klar.