typed+ir+codegen: Treat if expressions as expressions
Up until now we treated if expressions as statements. Now they are true
expressions. Instead of lowering if expressions to a form like this:

    if cond => a => else => b

    mut res
    if cond {
        res = a
    } else {
        res = b
    }
    res

we were able to solve the "lowering" much more elegantly during the IR
generation with just a few changes.

Furthermore, we now build and check anonymous unions more precisely.