asm: Add lifetimes to register allocations; Disable ir_remove_duplicate_getptr
This reduces the total amount of asm instructions of all examples in
`klar_the_book.md` from 10551 to 8378.

More work is needed as the register allocator does not correctly track
spilled registers over multiple blocks. Also, we want to get rid of the
conservative spill after a register write.

We had to disable the `ir_remove_duplicate_getptr` optimization because
it does not work with the new register allocator yet (because the latter
is flawed as stated above).